Abstract concurrency to separate class to make testing simpler and add tests
