+++ src/main/java/org/webscraper/Main.java
@@ -1,13 +1,16 @@
 package org.webscraper;
 
-
+import org.webscraper.concurrentservice.AsyncService;
 import org.webscraper.connector.ConnectorImpl;
+import org.webscraper.webscraper.WebScraper;
 
 public class Main {
     public static void main(String[] args) {
 
         new WebScraper(
-            new ConnectorImpl()
-        ).run("https://monzo.com");
+            new Scraper(new ConnectorImpl()),
+            new AsyncService(),
+            "https://monzo.com"
+        ).run();
     }
 }
