+++ src/main/java/org/webscraper/concurrentservice/ConcurrentService.java
@@ -1,0 +1,7 @@
+package org.webscraper.concurrentservice;
+
+public interface ConcurrentService {
+    void run(Runnable runnable);
+
+    void shutdown() ;
+}
